rename: Providers 파일 이름 수정 및 layout에서 가져오는 경로 수정
