+++ contexts/index.tsx
@@ -1,0 +1,1 @@
+export { default as Providers } from './Providers';
